feat(Global): start MarketingCloud SDK on startup
Instead of initializing the MC SDK when the init function is ran on the
JS layer the SDK should be initialized on the app startup (onCreate and
didFinishLaunchingWithOptions). That way the notification will more
reliable arrive when running on background
