+++ ios/react-native-sfmobilepush/SFMobilePush.h
@@ -4,7 +4,10 @@
 
 @interface SFMobilePush : RCTEventEmitter <RCTBridgeModule>
 
++ (void)initialize:(NSString *) applicationId accessToken:(NSString *) accessToken;
 + (void)didReceiveRemoteNotification:(UNNotificationResponse *) response;
++ (void)didRegisterForRemoteNotificationsWithDeviceToken:(NSData *)deviceToken;
++ (void)didReceiveRemoteNotificationUserInfo:(NSDictionary *) userInfo;
 - (void)handleRemoteNotification:(NSNotification *) notification;
 - (NSURL*) getConfigPath;
